pyproject.toml
[tool.poetry]
name = "try-elastic"
version = "0.1.0"
description = ""
authors = ["alina1412 <alina1412@mail.ru>"]
readme = "README.md"
packages = [{include = "try_elastic"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.19.0"
fastapi = "^0.87.0"
elasticsearch = "^8.5.0"
python-dotenv = "^0.21.0"
sqlalchemy = {extras = ["asyncio"], version = "^1.4.44"}
alembic = "^1.8.1"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
